closes https://github.com/datadryad/dryad-product-roadmap/issues/4321

+++ app/controllers/stash_engine/sessions_controller.rb
@@ -14,7 +14,7 @@ module StashEngine
 
     # this is the place omniauth calls back for shibboleth logins
     def callback
-      current_user.roles.tenant_roles.delete_all
+      current_user.roles.tenant_roles.delete_all if current_user.tenant_id != tenant.id
       current_user.update(tenant_id: params[:tenant_id], tenant_auth_date: Time.current)
       do_redirect
     end
@@ -130,7 +130,7 @@ module StashEngine
         redirect_to email_sso_path(tenant_id: current_user.email_token.tenant_id),
                     flash: { info: 'The code entered has expired. Check your email for a new code.' }
       elsif params[:token] == current_user.email_token.token
-        current_user.roles.tenant_roles.delete_all
+        current_user.roles.tenant_roles.delete_all if current_user.tenant_id != tenant.id
         current_user.update(tenant_id: current_user.email_token.tenant_id, tenant_auth_date: Time.current)
         do_redirect
       else
@@ -147,7 +147,7 @@ module StashEngine
         when 'email'
           redirect_to email_sso_path(tenant_id: tenant.id)
         when 'author_match'
-          current_user.roles.tenant_roles.delete_all
+          current_user.roles.tenant_roles.delete_all if current_user.tenant_id != tenant.id
           current_user.update(tenant_id: tenant.id, tenant_auth_date: Time.current)
           do_redirect
         when 'ip_address'
@@ -300,7 +300,7 @@ module StashEngine
         net = IPAddr.new(range)
         next unless net.include?(IPAddr.new(request.remote_ip))
 
-        current_user.roles.tenant_roles.delete_all
+        current_user.roles.tenant_roles.delete_all if current_user.tenant_id != tenant.id
         current_user.update(tenant_id: tenant.id, tenant_auth_date: Time.current)
         do_redirect
         return nil # adding nil here to jump out of loop and return early since rubocop sucks & requires a return value
@@ -325,7 +325,7 @@ module StashEngine
     def set_default_tenant
       return unless current_user.present?
 
-      current_user.roles.tenant_roles.delete_all
+      current_user.roles.tenant_roles.delete_all if current_user.tenant_id != APP_CONFIG.default_tenant
       current_user.update(tenant_id: APP_CONFIG.default_tenant)
     end
   end
